Only Inactive Walls
The abilities will only use inactive walls. However because of how it's done. You may see the already used walls restart their animations. This is because of our limit of vfx walls. When we use ability 1 then 2, 2 will use vfx 0 however 1 is already using vfx 0. So the vfx walls are the same and restart together