Shadow wall flash once only
What happened before was that when the enemy got into a wall once, the wall would always be desingated as having a player inside when destroyed or time ran out. Now whenever the exit wall function is called we always make sure to say the player is no longer inside.